Replace enemy state switch with a state-object machine
Every state is now an EnemyStateBase object owning its own data, driven by a reusable StateMachine<T>. The parallel override layer is gone: Grabbed, Ragdoll and Dead became states like the rest, so SetOverride, TickOverride and the nullable _override field all disappear, and Update/FixedUpdate reduce to one delegating line each. GrabbedState.Enter/Exit now guarantee the isKinematic toggle on every path out, including the break-out case the old code could not reach.

Adds Idle, Patrol, Searching, Chase, Attack, Grabbed, Ragdoll and Dead states. Chase, Searching and Idle are implemented; Patrol and Attack are still scaffolds. FollowPlayer was dropped as a duplicate of MoveToward, and _moveSpeed was renamed _chaseSpeed.

Player action states (Free/Firing/Holding) are scaffolded against the same StateMachine<T> and IState, ready for PlayerController to adopt. _bulletCount becomes an int for the six-round cylinder.

+++ Assets/Scripts/PlayerController.cs
@@ -4,7 +4,7 @@ using UnityEngine.InputSystem;
 public class PlayerController : MonoBehaviour
 {
     public float _moveSpeed;
-    public float _bulletCount;
+    public int _bulletCount;
     public GunController _gun;
     public bool _triggerHeld = false;
     private Rigidbody _myRigidbody;
